pyproject.toml
[project]
name = "cfgr"
version = "0.10.0"
description = "Config file manager and diff"
authors = [
    { name = "Eric Gustafson", email = "ericg@elfwerks.io" },
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.3",
    "PyYAML>=6.0.3",
    "pathspec>=1.1.0",
    "ydiff>=1.5",
]

[project.urls]
Homepage = "https://github.com/egustafson/cfgr"

[project.scripts]
cfgr = "cfgr:cli"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[tool.hatch.build.targets.wheel]
include = [
    "cfgr.py",
    "context.py",
    "filetree.py",
    "ops.py",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
